requirements.txt
datasets==3.6.0
evaluate==0.4.3
nltk==3.9.1
pandas==2.3.0
PyPDF2==3.0.1
python-dotenv==1.1.0
Requests==2.32.4
scikit_learn==1.7.0
spacy==3.8.7
streamlit==1.46.0
torch==2.7.1
transformers==4.52.4
